переносим логику в BoardController
