+++ src/controllers/board.js
@@ -1,9 +1,89 @@
+import {TaskCount} from "../utils/const";
+import {renderComponent, RenderPosition, replaceComponent, remove} from "../utils/render";
+
+import LoadMoreButtonComponent from "../components/load-more-button";
+import NoTasksComponent from "../components/no-task";
+import TaskEditComponent from "../components/task-edit";
+import SortComponent from "../components/sorting";
+import TaskComponent from "../components/task";
+import TasksComponent from "../components/tasks";
+
+const renderTask = (taskListElement, task) => {
+  const replaceTaskToEdit = () => {
+    replaceComponent(taskEditComponent, taskComponent);
+  };
+
+  const replaceEditToTask = () => {
+    replaceComponent(taskComponent, taskEditComponent);
+  };
+
+  const onEscKeyDown = (evt) => {
+    const isEscKey = evt.key === `Escape` || evt.key === `Esc`;
+
+    if (isEscKey) {
+      replaceEditToTask();
+      document.removeEventListener(`keydown`, onEscKeyDown);
+    }
+  };
+
+  const taskComponent = new TaskComponent(task);
+  const taskEditComponent = new TaskEditComponent(task);
+
+  taskComponent.setEditButtonClickHandler(() => {
+    replaceTaskToEdit();
+    document.addEventListener(`keydown`, onEscKeyDown);
+  });
+
+  taskEditComponent.setSubmitHandler((evt) => {
+    evt.preventDefault();
+    replaceEditToTask();
+    document.removeEventListener(`keydown`, onEscKeyDown);
+  });
+
+  renderComponent(taskListElement, taskComponent, RenderPosition.BEFOREEND);
+};
+
+const renderBoard = (boardComponent, tasks) => {
+  const isAllTasksArchived = tasks.every((task) => task.isArchive);
+
+  if (isAllTasksArchived) {
+    renderComponent(boardComponent.getElement(), new NoTasksComponent(), RenderPosition.BEFOREEND);
+    return;
+  }
+
+  renderComponent(boardComponent.getElement(), new SortComponent(), RenderPosition.BEFOREEND);
+  renderComponent(boardComponent.getElement(), new TasksComponent(), RenderPosition.BEFOREEND);
+
+  const taskListElement = boardComponent.getElement().querySelector(`.board__tasks`);
+
+  let showingTasksCount = TaskCount.SHOWING_ON_START;
+  tasks.slice(0, showingTasksCount)
+    .forEach((task) => {
+      renderTask(taskListElement, task);
+    });
+
+  const loadMoreButtonComponent = new LoadMoreButtonComponent();
+  renderComponent(boardComponent.getElement(), loadMoreButtonComponent, RenderPosition.BEFOREEND);
+
+  loadMoreButtonComponent.setClickHandler(() => {
+    const prevTasksCount = showingTasksCount;
+    showingTasksCount = showingTasksCount + TaskCount.SHOWING_BY_BUTTON;
+
+    tasks.slice(prevTasksCount, showingTasksCount)
+      .forEach((task) => renderTask(taskListElement, task));
+
+    if (showingTasksCount >= tasks.length) {
+      remove(loadMoreButtonComponent);
+    }
+  });
+};
+
 export default class BoardController {
   constructor(container) {
-
+    this._container = container;
   }
 
   render(tasks) {
-
+    renderBoard(this._container, tasks);
   }
 }
